Added images for podcasts, made reusable ResourceList component to use in podcasts and learn pages
